Slew of bug fixes
