+++ blues_refactor/simulation.py
@@ -1,0 +1,266 @@
+from __future__ import print_function
+import simtk.unit as unit
+from alchemy import AbsoluteAlchemicalFactory, AlchemicalState
+import numpy as np
+
+from simtk import unit, openmm
+from simtk.openmm import app
+from alchemy import AbsoluteAlchemicalFactory, AlchemicalState
+
+import utils
+from blues.ncmc_switching import NCMCVVAlchemicalIntegrator
+
+import sys, parmed, math, copy
+import numpy as np
+import mdtraj
+from mdtraj.reporters import HDF5Reporter
+from datetime import datetime
+from optparse import OptionParser
+
+class Simulation(object):
+    """Simulation class provides the functions that perform the BLUES run.
+
+    Ex.
+        import blues.ncmc
+        blues = ncmc.Simulation(sims, model, mover, **opt)
+        blues.run()
+
+    """
+    def __init__(self, simulations, model, mover, **opt):
+        """Initialize the BLUES Simulation object.
+
+        Parameters
+        ----------
+        sims : blues.ncmc.SimulationFactory object
+            SimulationFactory Object which carries the 3 required
+            OpenMM Simulation objects (MD, NCMC, ALCH) required to run BLUES.
+        model : blues.ncmc.ModelProperties object
+            ModelProperties object that represents the model to be perturbed
+            in the NCMC simulation.
+        mover : blues.ncmc.MoveProposal object
+            MoveProposal object which contains the dict of moves performed
+            in the NCMC simulation.
+        """
+        self.md_sim = simulations.md
+        self.alch_sim = simulations.alch
+        self.nc_context = simulations.nc.context
+        self.nc_integrator = simulations.nc.context._integrator
+        self.model = model
+        self.mover = mover
+
+        self.accept = 0
+        self.reject = 0
+        self.accept_ratio = 0
+
+        self.nIter = int(opt['nIter'])
+        self.nstepsNC = int(opt['nstepsNC'])
+        self.nstepsMD = int(opt['nstepsMD'])
+
+        self.current_iter = 0
+        self.current_stepNC = 0
+        self.current_stepMD = 0
+
+        self.current_state = { 'md'   : { 'state0' : {}, 'state1' : {} },
+                               'nc'   : { 'state0' : {}, 'state1' : {} },
+                               'alch' : { 'state0' : {}, 'state1' : {} }
+                            }
+
+        self.temperature = opt['temperature']
+        kB = unit.BOLTZMANN_CONSTANT_kB * unit.AVOGADRO_CONSTANT_NA
+        kT = kB * opt['temperature']
+        beta = 1.0 / kT
+        self.beta = beta
+
+        self.work_keys = ['total_work', 'lambda', 'shadow_work',
+                          'protocol_work', 'Eold', 'Enew','Epert']
+
+        self.state_keys = { 'getPositions' : True,
+                       'getVelocities' : True,
+                       'getForces' : False,
+                       'getEnergy' : True,
+                       'getParameters': True,
+                       'enforcePeriodicBox' : True}
+
+    def setSimState(self, simkey, stateidx, stateinfo):
+        """Stores the dict of Positions, Velocities, Potential/Kinetic energies
+        of the state before and after a NCMC step or iteration.
+
+        Parameters
+        ----------
+        simkey : str (key: 'md', 'nc', 'alch')
+            Key corresponding to the simulation.
+        stateidx : int (key: 'state0' or 'state1')
+            Key corresponding to the state information being stored.
+        stateinfo : dict
+            Dictionary containing the State information.
+        """
+        self.current_state[simkey][stateidx] = stateinfo
+
+    def setStateConditions(self):
+        """Stores the dict of current state of the MD and NCMC simulations.
+        Dict contains the Positions, Velocities, Potential/Kinetic Energies
+        of the current state.
+        Sets the NCMC simulation Positions/Velocities to
+        the current state of the MD simulation.
+        """
+        md_state0 = self.getStateInfo(self.md_sim.context, self.state_keys)
+        nc_state0 = self.getStateInfo(self.nc_context, self.state_keys)
+        self.nc_context.setPositions(md_state0['positions'])
+        self.nc_context.setVelocities(md_state0['velocities'])
+        self.setSimState('md', 'state0', md_state0)
+        self.setSimState('nc', 'state0', nc_state0)
+
+    def getStateInfo(self, context, parameters):
+        """Function that gets the State information from the given context and
+        list of parameters to queuey it with.
+        Returns a dict of the data from the State.
+
+        Parameters
+        ----------
+        context : openmm.Context
+            Context of the OpenMM Simulation to query.
+        parameters : list
+            Default: [ positions, velocities, potential_energy, kinetic_energy ]
+            A list that defines what information to get from the context State.
+        """
+        stateinfo = {}
+        state  = context.getState(**parameters)
+        stateinfo['iter'] = int(self.current_iter)
+        stateinfo['positions'] =  state.getPositions(asNumpy=True)
+        stateinfo['velocities'] = state.getVelocities(asNumpy=True)
+        stateinfo['potential_energy'] = state.getPotentialEnergy()
+        stateinfo['kinetic_energy'] = state.getKineticEnergy()
+        return stateinfo
+
+    def getWorkInfo(self, nc_integrator, parameters):
+        """Function that obtains the work and energies from the NCMC integrator.
+
+        Returns a dict of the specified parameters.
+
+        Parameters
+        ----------
+        nc_integrator : openmm.Context.Integrator
+            The integrator from the NCMC Context
+        parameters : list
+            list containing strings of the values to get from the integrator.
+            Default : ['total_work', 'lambda', 'shadow_work',
+                       'protocol_work', 'Eold', 'Enew','Epert']
+        """
+        workinfo = {}
+        for param in parameters:
+            workinfo['param'] = nc_integrator.getGlobalVariableByName(param)
+        return workinfo
+
+    def chooseMove(self):
+        """Function that chooses to accept or reject the proposed move.
+        """
+        md_state0 = self.current_state['md']['state0']
+        nc_state0 = self.current_state['nc']['state0']
+        nc_state1 = self.current_state['nc']['state1']
+
+        log_ncmc = self.nc_integrator.getLogAcceptanceProbability(self.nc_context)
+        randnum =  math.log(np.random.random())
+
+        ### Compute Alchemical Correction Term
+        if np.isnan(log_ncmc) == False:
+            self.alch_sim.context.setPositions(nc_state1['positions'])
+            alch_state1 = self.getStateInfo(self.alch_sim.context, self.state_keys)
+            self.setSimState('alch', 'state1', alch_state1)
+
+            #print('alc_oldPE', nc_state0['potential_energy'])
+            #print('oldPE', md_state0['potential_energy'] )
+            #print('norm_newPE', nc_state1['potential_energy'])
+            #print('alc_newPE', alch_state1['potential_energy'])
+
+            correction_factor = (nc_state0['potential_energy'] - md_state0['potential_energy'] + nc_state1['potential_energy'] - alch_state1['potential_energy']) * (1.0/self.nc_integrator.kT)
+            # Should this be negative 1?
+            #n1_PE = alch_state1['potential_energy'] - nc_state1['potential_energy']
+            #n_PE = md_state0['potential_energy'] - nc_state0['potential_energy']
+            #correction_factor = (-1.0/self.nc_integrator.kT)*( n1_PE - n_PE )
+            #print('correction_factor', correction_factor)
+            log_ncmc = log_ncmc + correction_factor
+
+        if log_ncmc > randnum:
+            self.accept += 1
+            print('NCMC MOVE ACCEPTED: log_ncmc {} > randnum {}'.format(log_ncmc, randnum) )
+            #print('accCounter', float(self.accept)/float(stepsdone+1), self.accept)
+            self.md_sim.context.setPositions(nc_state1['positions'])
+        else:
+            self.reject += 1
+            print('NCMC MOVE REJECTED: {} < {}'.format(log_ncmc, randnum) )
+            #print('ncmc PE', newinfo['potential_energy'], 'old PE', md_PE0)
+            self.nc_context.setPositions(md_state0['positions'])
+
+        self.nc_integrator.reset()
+        self.md_sim.context.setVelocitiesToTemperature(self.temperature)
+
+    def simulateNCMC(self):
+        """Function that performs the NCMC simulation."""
+        for nc_step in range(self.nstepsNC):
+            try:
+                self.current_stepNC = int(nc_step)
+                # Calculate Work/Energies Before Step
+                work_initial = self.getWorkInfo(self.nc_integrator, self.work_keys)
+
+                # Attempt NCMC Move
+                if int(self.mover.moves['step']) == nc_step:
+                    print('[Iter {}] Performing NCMC {} move'.format(
+                    self.current_iter, self.mover.moves['method'].__name__))
+
+                    #Do move
+                    self.model, self.nc_context = self.mover.moves['method'](model=self.model, nc_context=self.nc_context)
+
+                # Do 1 NCMC step
+                self.nc_integrator.step(1)
+                # Calculate Work/Energies After Step.
+                work_final = self.getWorkInfo(self.nc_integrator, self.work_keys)
+            except Exception as e:
+                print(e)
+                break #should this be exit?
+
+        nc_state1 = self.getStateInfo(self.nc_context, self.state_keys)
+        self.setSimState('nc', 'state1', nc_state1)
+
+    def simulateMD(self):
+        """Function that performs the MD simulation."""
+        md_state0 = self.current_state['md']['state0']
+        try:
+            self.md_sim.step(self.nstepsMD)
+            self.current_stepMD = self.md_sim.currentStep
+        except Exception as e:
+            print(e)
+            stateinfo = self.getStateInfo(self.md_sim.context, self.state_keys)
+            last_x, last_y = np.shape(md_state0['positions'])
+            reshape = (np.reshape(md_state0['positions'], (1, last_x, last_y))).value_in_unit(unit.nanometers)
+            print('potential energy before NCMC', md_state0['potential_energy'])
+            print('kinetic energy before NCMC', md_state0['kinetic_energy'])
+
+            last_top = mdtraj.Topology.from_openmm(self.md_sim.topology)
+            broken_frame = mdtraj.Trajectory(xyz=reshape, topology=last_top)
+            broken_frame.save_pdb('MD-blues_fail-iter{}_md{}.pdb'.format(self.current_iter, self.current_stepMD))
+            exit()
+
+        md_state1 = self.getStateInfo(self.md_sim.context, self.state_keys)
+        self.setSimState('md', 'state1', md_state1)
+        # Set NC poistions to last positions from MD
+        self.nc_context.setPositions(md_state1['positions'])
+        self.nc_context.setVelocities(md_state1['velocities'])
+
+    def run(self):
+        """Function that runs the BLUES engine that iterates of the actions:
+        Perform NCMC simulation, perform proposed move, accepts/rejects move,
+        then performs the MD simulation from the NCMC state.
+        """
+        #set inital conditions
+        self.setStateConditions()
+        for n in range(self.nIter):
+            self.current_iter = int(n)
+            self.setStateConditions()
+            self.simulateNCMC()
+            self.chooseMove()
+            self.simulateMD()
+
+        # END OF NITER
+        self.accept_ratio = self.accept/float(self.nIter)
+        print('Acceptance Ratio', self.accept_ratio)
+        print('numsteps ', self.nstepsNC)
